+++ src/main/kotlin/net/ccbluex/liquidbounce/web/theme/ThemeManager.kt
@@ -33,7 +33,9 @@ import net.ccbluex.liquidbounce.web.browser.supports.tab.ITab
 import net.ccbluex.liquidbounce.web.integration.IntegrationHandler
 import net.ccbluex.liquidbounce.web.integration.VirtualScreenType
 import net.ccbluex.liquidbounce.web.socket.netty.NettyServer.Companion.NETTY_ROOT
+import net.ccbluex.liquidbounce.web.theme.component.Component
 import net.ccbluex.liquidbounce.web.theme.component.ComponentOverlay
+import net.ccbluex.liquidbounce.web.theme.component.ComponentType
 import net.minecraft.client.gui.screen.ChatScreen
 import java.io.File
 
@@ -52,7 +54,7 @@ object ThemeManager {
             field = value
 
             // Update components
-            ComponentOverlay.parseComponents()
+            ComponentOverlay.insertComponents()
 
             // Update integration browser
             IntegrationHandler.updateIntegrationBrowser()
@@ -110,6 +112,8 @@ class Theme(val name: String) {
     private val url: String
         get() = "$NETTY_ROOT/$name/#/"
 
+
+
     /**
      * Get the URL to the given page name in the theme.
      */
@@ -126,6 +130,33 @@ class Theme(val name: String) {
     fun doesSupport(name: String?) = name != null && metadata.supports.contains(name)
 
     fun doesOverlay(name: String?) = name != null && metadata.overlays.contains(name)
+
+    fun parseComponents(): MutableList<Component> {
+        val themeComponent = metadata.rawComponents
+            .map { it.asJsonObject }
+            .associateBy { it["name"].asString!! }
+
+        val componentList = mutableListOf<Component>()
+
+        for ((name, obj) in themeComponent) {
+            runCatching {
+                val componentType = ComponentType.byName(name) ?: error("Unknown component type: $name")
+                val component = componentType.createComponent()
+
+                runCatching {
+                    ConfigSystem.deserializeConfigurable(component, obj)
+                }.onFailure {
+                    logger.error("Failed to deserialize component $name", it)
+                }
+
+                componentList.add(component)
+            }.onFailure {
+                logger.error("Failed to create component $name", it)
+            }
+        }
+
+        return componentList
+    }
 
     companion object {
